requirements.txt
streamlit
google-generativeai
sqlite3 
requests
yagmail
hashlib
python-dotenv

